requirements.txt
python-docx
Flask
docxcompose
docx2pdf